+++ src/service.rs
@@ -1638,6 +1638,27 @@ impl Service {
         }
     }
 
+    /// Set service description.
+    ///
+    /// Required permission: [`ServiceAccess::CHANGE_CONFIG`].
+    pub fn set_description(&self, description: impl AsRef<OsStr>) -> crate::Result<()> {
+        let wide_str =
+            WideCString::from_os_str(description).map_err(Error::InvalidServiceDescription)?;
+        let mut wide_str_vec = wide_str.into_vec_with_nul();
+
+        let mut service_description_request =
+            unsafe { mem::zeroed::<winsvc::SERVICE_DESCRIPTIONW>() };
+        service_description_request.lpDescription = wide_str_vec.as_mut_ptr();
+
+        unsafe {
+            self.change_config2(
+                winsvc::SERVICE_CONFIG_DESCRIPTION,
+                &mut service_description_request,
+            )
+            .map_err(Error::Winapi)
+        }
+    }
+
     /// Private helper to send the control commands to the system.
     fn send_control_command(&self, command: ServiceControl) -> crate::Result<ServiceStatus> {
         let mut raw_status = unsafe { mem::zeroed::<winsvc::SERVICE_STATUS>() };